+++ src/Column.php
@@ -24,8 +24,7 @@ class Column
     public $htmlClosure;
     public $icon;
     public $displayIconWhenNoValue;
-    public $columnClasses;
-    public $resultClosures;
+    public $classes;
 
     /**
      * \Okipa\LaravelTable\Column constructor.
@@ -39,7 +38,6 @@ class Column
         $this->databaseDefaultTable = $table->model->getTable();
         $this->databaseDefaultColumn = $databaseColumn;
         $this->title = $databaseColumn ? __('validation.attributes.' . $databaseColumn) : null;
-        $this->resultClosures = new Collection();
     }
 
     /**
@@ -240,24 +238,8 @@ class Column
      */
     public function classes(array $classes): Column
     {
-        $this->columnClasses = $classes;
+        $this->classes = $classes;
 
-        return $this;
-    }
-
-    /**
-     * Set a result output that will be displayed under the column (html accepted).
-     * The closure let you manipulate the following attributes : $displayedList.
-     * You can call this method as much time as you need on a column to append the wished number of results.
-     *
-     * @param \Closure $resultClosure
-     *
-     * @return \Okipa\LaravelTable\Column
-     */
-    public function result(Closure $resultClosure): Column
-    {
-        $this->resultClosures->push($resultClosure);
-        
         return $this;
     }
 }