+++ tests/Unit/ResultDeclarationTest.php
@@ -7,16 +7,14 @@ use Okipa\LaravelTable\Test\LaravelTableTestCase;
 use Okipa\LaravelTable\Test\Models\Company;
 use Okipa\LaravelTable\Test\Models\User;
 
-class ResultsTest extends LaravelTableTestCase
+class ResultDeclarationTest extends LaravelTableTestCase
 {
     public function testSetResultsAttribute()
     {
-        $closure = function ($displayedList) {
-            $displayedList->test = 'hello';
-        };
         $table = (new Table)->model(User::class);
-        $table->column()->result($closure);
-        $this->assertEquals($closure, $table->columns->first()->resultClosures->first());
+        $table->result()->title('Test');
+        $this->assertEquals($table->results->count(), 1);
+        $this->assertEquals($table->results->first()->title, 'Test');
     }
 
     public function testSetResultsHtml()
@@ -24,15 +22,15 @@ class ResultsTest extends LaravelTableTestCase
         $this->createMultipleUsers(10);
         $companies = $this->createMultipleCompanies(5);
         $this->routes(['users'], ['index']);
-        $table = (new Table)->model(Company::class)
-            ->routes(['index' => ['name' => 'users.index']]);
+        $table = (new Table)->model(Company::class)->routes(['index' => ['name' => 'users.index']]);
         $table->column('name');
-        $table->column('turnover')->result(function ($displayedList) {
+        $table->column('turnover');
+        $table->result()->title('Result !')->html(function ($displayedList) {
             return $displayedList->sum('turnover');
         });
         $table->render();
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
-        $this->assertEquals([null, $companies->sum('turnover')], $table->results->first()->toArray());
+        $this->assertContains('Result !', $html);
         $this->assertContains((string) $companies->sum('turnover'), $html);
     }
 
@@ -45,15 +43,18 @@ class ResultsTest extends LaravelTableTestCase
             ->routes(['index' => ['name' => 'users.index']])
             ->rowsNumber(2);
         $table->column('name');
-        $table->column('turnover')->result(function ($displayedList) {
+        $table->column('turnover');
+        $table->result()->title('Selected turnover')->html(function ($displayedList) {
             return $displayedList->sum('turnover');
-        })->result(function () {
+        });
+        $table->result()->title('Total turnover')->html(function () {
             return (new Company)->all()->sum('turnover');
         });
         $table->render();
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
-        $this->assertContains((string) $table->list->getCollection()->sum('turnover'), $html);
+        $this->assertContains('Selected turnover', $html);
         $this->assertContains((string) $companies->sum('turnover'), $html);
+        $this->assertContains('Total turnover', $html);
     }
 
     public function testSetNoResult()
@@ -68,7 +69,6 @@ class ResultsTest extends LaravelTableTestCase
         $table->column('turnover');
         $table->render();
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
-        $this->assertNotContains('results', $html);
         $this->assertNotContains('result', $html);
     }
 }